requirements.txt
streamlit>=1.33.0
psycopg2-binary
pandas
fpdf2>=2.7.0